pyproject.toml
[project]
name = "rendering-testing-for-backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.110)",
    "uvicorn (>=0.29)",
    "gunicorn (>=21.2)",
    "python-dotenv (>=1.0)",
    "kiteconnect (>=4.2)",
    "requests (>=2.31)",
    "passlib[argon2] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]
